Tidy up signing test code
* Move signing tests to mariner2_test and windows_test, respectively,
since the test is checking 1) that the forwarding happens and 2) that
the signing frontend receives sufficient information to determine
signing logic.
* Simplify the test signing frontend to do less.
