+++ test/mariner2_test.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"os"
+	"strings"
 	"testing"
 
 	"github.com/Azure/dalec"
@@ -87,6 +88,7 @@ func testLinuxDistro(ctx context.Context, t *testing.T, buildTarget string) {
 			return gwclient.NewResult(), nil
 		})
 	})
+
 	t.Run("container", func(t *testing.T) {
 		spec := dalec.Spec{
 			Name:        "test-container-build",
@@ -299,6 +301,74 @@ echo "$BAR" > bar.txt
 
 			if outErr != nil {
 				return nil, outErr
+			}
+
+			return gwclient.NewResult(), nil
+		})
+	})
+
+	runTest := func(t *testing.T, f gwclient.BuildFunc) {
+		t.Helper()
+		ctx := startTestSpan(baseCtx, t)
+		testEnv.RunTest(ctx, t, f)
+	}
+
+	t.Run("test mariner2 signing", func(t *testing.T) {
+		t.Parallel()
+		runTest(t, func(ctx context.Context, gwc gwclient.Client) (*gwclient.Result, error) {
+			spec := dalec.Spec{
+				Name:        "foo",
+				Version:     "v0.0.1",
+				Description: "foo bar baz",
+				Website:     "https://foo.bar.baz",
+				Revision:    "1",
+				Targets: map[string]dalec.Target{
+					"mariner2": {
+						PackageConfig: &dalec.PackageConfig{
+							Signer: &dalec.Frontend{
+								Image: phonySignerRef,
+							},
+						},
+					},
+				},
+				Sources: map[string]dalec.Source{
+					"foo": {
+						Inline: &dalec.SourceInline{
+							File: &dalec.SourceInlineFile{
+								Contents: "#!/usr/bin/env bash\necho \"hello, world!\"\n",
+							},
+						},
+					},
+				},
+				Build: dalec.ArtifactBuild{
+					Steps: []dalec.BuildStep{
+						{
+							Command: "/bin/true",
+						},
+					},
+				},
+				Artifacts: dalec.Artifacts{
+					Binaries: map[string]dalec.ArtifactConfig{
+						"foo": {},
+					},
+				},
+			}
+
+			sr := newSolveRequest(withSpec(ctx, t, &spec), withBuildTarget("mariner2/rpm"))
+			res, err := gwc.Solve(ctx, sr)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			tgt := readFile(ctx, t, "/target", res)
+			cfg := readFile(ctx, t, "/config.json", res)
+
+			if string(tgt) != "mariner2" {
+				t.Fatal(fmt.Errorf("target incorrect; either not sent to signer or not received back from signer"))
+			}
+
+			if !strings.Contains(string(cfg), "linux") {
+				t.Fatal(fmt.Errorf("configuration incorrect"))
 			}
 
 			return gwclient.NewResult(), nil
